fix: dead click fixes from watching in prod
