+++ src/autocapture-utils.ts
@@ -4,6 +4,7 @@ import { each, entries, includes, trim } from './utils'
 import { isArray, isNullish, isString, isUndefined } from './utils/type-utils'
 import { logger } from './utils/logger'
 import { window } from './utils/globals'
+import { isDocumentFragment, isElementNode, isTag, isTextNode } from './utils/element-utils'
 
 export function splitClassString(s: string): string[] {
     return s ? trim(s).split(/\s+/) : []
@@ -92,47 +93,6 @@ export function getEventTarget(e: Event): Element | null {
         }
         return (e.target as Element) || null
     }
-}
-
-/*
- * Check whether an element has nodeType Node.ELEMENT_NODE
- * @param {Element} el - element to check
- * @returns {boolean} whether el is of the correct nodeType
- */
-export function isElementNode(el: Node | Element | undefined | null): el is Element {
-    return !!el && el.nodeType === 1 // Node.ELEMENT_NODE - use integer constant for browser portability
-}
-
-/*
- * Check whether an element is of a given tag type.
- * Due to potential reference discrepancies (such as the webcomponents.js polyfill),
- * we want to match tagNames instead of specific references because something like
- * element === document.body won't always work because element might not be a native
- * element.
- * @param {Element} el - element to check
- * @param {string} tag - tag name (e.g., "div")
- * @returns {boolean} whether el is of the given tag type
- */
-export function isTag(el: Element | undefined | null, tag: string): el is HTMLElement {
-    return !!el && !!el.tagName && el.tagName.toLowerCase() === tag.toLowerCase()
-}
-
-/*
- * Check whether an element has nodeType Node.TEXT_NODE
- * @param {Element} el - element to check
- * @returns {boolean} whether el is of the correct nodeType
- */
-export function isTextNode(el: Element | undefined | null): el is HTMLElement {
-    return !!el && el.nodeType === 3 // Node.TEXT_NODE - use integer constant for browser portability
-}
-
-/*
- * Check whether an element has nodeType Node.DOCUMENT_FRAGMENT_NODE
- * @param {Element} el - element to check
- * @returns {boolean} whether el is of the correct nodeType
- */
-export function isDocumentFragment(el: Element | ParentNode | undefined | null): el is DocumentFragment {
-    return !!el && el.nodeType === 11 // Node.DOCUMENT_FRAGMENT_NODE - use integer constant for browser portability
 }
 
 export const autocaptureCompatibleElements = ['a', 'button', 'form', 'input', 'select', 'textarea', 'label']
